DTSPO-6161 Upgrade app insights (#30)
