+++ .gitignore
@@ -18,3 +18,6 @@
 
 .DS_store
 out/
+
+applicationinsights-agent-*.jar
+*.log
