+++ Dockerfile
@@ -1,7 +1,6 @@
-ARG APP_INSIGHTS_AGENT_VERSION=2.6.4
+ARG APP_INSIGHTS_AGENT_VERSION=3.2.4
 FROM hmctspublic.azurecr.io/base/java:17-distroless
 
-COPY lib/AI-Agent.xml /opt/app/
 COPY build/libs/sds-toffee-recipes-service.jar /opt/app/
 
 CMD ["sds-toffee-recipes-service.jar"]
