+++ build.gradle
@@ -88,10 +88,6 @@ dependencyUpdates {
 
 ext["rest-assured.version"] = '4.4.0'
 
-ext {
-  loggingVersion = "5.1.9"
-}
-
 dependencies {
   implementation group: 'org.springframework.boot', name: 'spring-boot-starter-web'
   implementation group: 'org.springframework.boot', name: 'spring-boot-starter-actuator'
@@ -102,8 +98,7 @@ dependencies {
 
   implementation group: 'com.google.guava', name: 'guava', version: '31.0.1-jre'
 
-  implementation group: 'com.github.hmcts.java-logging', name: 'logging', version: loggingVersion
-  implementation group: 'com.github.hmcts.java-logging', name: 'logging-appinsights', version: loggingVersion
+  implementation group: 'com.github.hmcts.java-logging', name: 'logging', version: '5.1.9'
   implementation group: 'org.springdoc', name: 'springdoc-openapi-ui', version: '1.6.4'
 
   compileOnly group: 'org.projectlombok', name: 'lombok', version: lombokVersion
@@ -141,8 +136,8 @@ sourceSets {
 }
 
 configurations {
-  functionalCompile.extendsFrom testImplementation
-  functionalRuntime.extendsFrom testRuntime
+  functionalImplementation.extendsFrom testImplementation
+  functionalRuntimeOnly.extendsFrom testRuntimeOnly
 }
 
 task highLevelDataSetup {
@@ -153,25 +148,22 @@ task smoke(type: Test, description: 'Runs the smoke tests.', group: 'Verificatio
   testClassesDirs = sourceSets.functional.output.classesDirs
   classpath = sourceSets.functional.runtimeClasspath
 
-  useJUnit {
-    includeCategories 'uk.gov.hmcts.reform.rhubarb.functional.SmokeTest'
+  useJUnitPlatform {
+    includeTags "SmokeTest"
   }
 }
 
 task functional(type: Test, description: 'Runs the functional tests.', group: 'Verification') {
   testClassesDirs = sourceSets.functional.output.classesDirs
   classpath = sourceSets.functional.runtimeClasspath
+
+  useJUnitPlatform()
 }
 
 test {
   useJUnitPlatform()
 }
 
-tasks.withType(Test) {
-  useJUnitPlatform()
-}
-
 wrapper {
   distributionType = Wrapper.DistributionType.ALL
 }
-
