--- charts/apple-recipe-backend/values.aat.template.yaml
@@ -1,6 +1,0 @@
-java:
-  # Don't modify below here
-  image: ${IMAGE_NAME}
-  ingressHost: ${SERVICE_FQDN}
-  environment:
-    POSTGRES_SSL_MODE: require
