--- charts/apple-recipe-backend/values.preview.template.yaml
@@ -1,7 +1,0 @@
-java:
-  # Don't modify below here
-  image: ${IMAGE_NAME}
-  ingressHost: ${SERVICE_FQDN}
-  postgresql:
-    enabled: true
-  disableKeyVaults: true
