+++ charts/apple-recipe-backend/values.yaml
@@ -1,7 +1,10 @@
 java:
   applicationPort: 4550
-  # Don't modify below here
   image: 'hmctssandbox.azurecr.io/hmcts/apple-recipe-backend:latest'
+  aadIdentityName: apple
+  environment:
+    POSTGRES_SSL_MODE: require
+    APPLICATIONINSIGHTS_CONFIGURATION_FILE: /mnt/secrets/applesi/app-insights-config
   keyVaults:
     "applesi":
       secrets:
@@ -13,3 +16,5 @@ java:
           alias: POSTGRES_USER
         - name: recipe-backend-POSTGRES-PASS
           alias: POSTGRES_PASSWORD
+        - name: recipe-backend-app-insights-config
+          alias: app-insights-config
