+++ charts/toffee-recipe-backend/Chart.yaml
@@ -1,7 +1,7 @@
 name: toffee-recipe-backend
 apiVersion: v2
 home: https://github.com/hmcts/sds-toffee-recipes-service
-version: 0.8.7
+version: 0.8.8
 description: HMCTS toffee recipes service (test application)
 maintainers:
   - name: HMCTS Platform Engineering Team
