+++ infrastructure/output.tf
@@ -1,12 +1,4 @@
-output "app_namespace" {
-  value = var.deployment_namespace
-}
-
 # this variable will be accessible to tests as API_GATEWAY_URL environment variable
 output "api_gateway_url" {
   value = "https://core-api-mgmt-${var.env}.azure-api.net/${local.api_base_path}"
-}
-
-output "vaultName" {
-  value = local.vault_name
 }
