--- lib/AI-Agent.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<ApplicationInsightsAgent>
-    <AgentLogger>INFO</AgentLogger>
-    <Instrumentation>
-        <BuiltIn enabled="true">
-            <HTTP enabled="true"/>
-            <JDBC enabled="true"/>
-        </BuiltIn>
-    </Instrumentation>
-</ApplicationInsightsAgent>
