+++ lib/applicationinsights.json
@@ -1,0 +1,5 @@
+{
+  "role": {
+    "name": "sds-toffee-recipes-service"
+  }
+}
