--- src/test/functional/java/uk/gov/hmcts/reform/rhubarb/functional/SmokeTest.java
@@ -1,3 +1,0 @@
-package uk.gov.hmcts.reform.rhubarb.functional;
-
-public interface SmokeTest { /* category marker */ }
